Use relative imports in fluent-web and fluent-dom (#58)
Fixes #57.

Both fluent-web and fluent-dom are at versions 0.0.1. I expect the API
to fluctuate before it stabilizes. Let's use relative imports for the
while being so that it's easier to iterate.

Once we reach 0.1, we can bring back package imports and proper
dependency management via package.json.

Along with this, this PR cleans up a few things:

  - I renamed `resIds` to `resourceIds` in line with the change from
    fb50012

  - I removed `CachedIterable` from fluent-web; as of 5f45963 it is part
    of fluent-dom.